To be continued

+++ src/tank.rs
@@ -1,17 +1,17 @@
 pub struct Tank {
-    pub x: f32,
-    pub y: f32,
-    pub vx: f32,
-    pub vy: f32,
+    pub x: u32,
+    pub y: u32,
+    pub vx: u32,
+    pub vy: u32,
 }
 
 impl Tank {
     pub fn new() -> Tank {
         Tank {
-            x: 0.0,
-            y: 0.0,
-            vx: 0.0,
-            vy: 0.0,
+            x: 0,
+            y: 0,
+            vx: 0,
+            vy: 0,
         }
     }
 }
